requirements.txt
Flask==0.10.1
kafka-python==0.9.2
Flask-UUID
gevent
uwsgi
Flask-uWSGI-WebSocket
sphinx


